+++ src/game/tests.rs
@@ -432,6 +432,7 @@ mod vintage_cube_memory_lapse;
 mod vintage_cube_metamorphosis_fanatic;
 mod vintage_cube_mine_collapse;
 mod vintage_cube_monarch;
+mod vintage_cube_monstrous_rage;
 mod vintage_cube_more_spells;
 mod vintage_cube_mox_opal;
 mod vintage_cube_multiversal_passage;